% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_files.R
\name{check_files_exist}
\alias{check_files_exist}
\title{Check that files exist}
\usage{
check_files_exist(files_vec, error_file_prefix = "File")
}
\arguments{
\item{files_vec}{A vector of file paths}
}
\value{
TRUE if all input bedfiles have an associated tabix index file.
FALSE if not
}
\description{
Check that files exist
}
