% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cpp_make_bsseq}
\alias{Cpp_make_bsseq}
\title{Make a Bsseq object}
\usage{
Cpp_make_bsseq(bedfiles, regions, nthreads)
}
\arguments{
\item{bedfiles}{A vector of bedfiles}

\item{regions}{A vector of regions}
}
\description{
Make a Bsseq object
}
\keyword{internal}
