% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cpp_query_all_reference}
\alias{Cpp_query_all_reference}
\title{Make the inputs to a BSseq object
Produces M and coverage matrices, 'chr:start' of the loci as rownames,
sample names}
\usage{
Cpp_query_all_reference(
  bedfiles,
  regions,
  mmat,
  cmat,
  bismark,
  merged,
  sparse,
  nthreads
)
}
\arguments{
\item{bedfiles}{A vector of bedfiles}

\item{regions}{A vector of regions}

\item{mmat}{A NumericMatrix from R for M values - must be large enough to
hold the final matrix produced or a copy will be made}

\item{cmat}{A NumericMatrix from R for coverage - must be large enough to
hold the final matrix produced or a copy will be made}

\item{bismark}{Whether the input is a bismark coverage file}

\item{nthreads}{Set number of threads to use overriding the
\code{"iscream.threads"} option. See \code{?set_threads} for more information.}
}
\description{
Make the inputs to a BSseq object
Produces M and coverage matrices, 'chr:start' of the loci as rownames,
sample names
}
