% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cpp_query_all_sparse}
\alias{Cpp_query_all_sparse}
\title{Make the inputs to a BSseq object
Produces M and coverage matrices, 'chr:start' of the loci as rownames, sample names}
\usage{
Cpp_query_all_sparse(bedfiles, regions, bismark, merged, sparse, nthreads)
}
\arguments{
\item{bedfiles}{A vector of bedfiles}

\item{regions}{A vector of regions}

\item{bismark}{Whether the input is a bismark coverage file}

\item{nthreads}{Set number of threads to use overriding the
\code{"iscream.threads"} option. See \code{?set_threads} for more information.}
}
\description{
Make the inputs to a BSseq object
Produces M and coverage matrices, 'chr:start' of the loci as rownames, sample names
}
