% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cpp_query_chroms}
\alias{Cpp_query_chroms}
\title{Query the chromosomes or seqnames from a vector of files}
\usage{
Cpp_query_chroms(bedfile_vec, nthreads)
}
\arguments{
\item{bedfile_vec}{The vector of bedfile paths}
}
\value{
A vector of seqnames
}
\description{
Query the chromosomes or seqnames from a vector of files
}
\keyword{internal}
