% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cpp_query_interval}
\alias{Cpp_query_interval}
\title{Get reads from a single genomic region from one tabixed bed file to return as CharacterVector}
\usage{
Cpp_query_interval(bedfile, regions)
}
\arguments{
\item{bedfile}{The name of the bed file - must have a corresponding tabix
file with the same name and .tbi extension}

\item{regions}{A vector of region strings in the form "chr:start-end"}
}
\description{
Get reads from a single genomic region from one tabixed bed file to return as CharacterVector
}
