% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cpp_set_log_level}
\alias{Cpp_set_log_level}
\title{spdlog Logging Lever Setter}
\usage{
Cpp_set_log_level(name)
}
\arguments{
\item{name}{A string with the logging level. Value understood are,
in decreasing verbosity \sQuote{trace}, \sQuote{debug}, \sQuote{info},
\sQuote{warning}, \sQuote{error}, \sQuote{critical}, and \sQuote{off}.
Unrecognised names are equivalent to \sQuote{off}.}
}
\value{
Nothing is returned.
}
\description{
A helper function to turn a logging level given as string
into the current logging level
}
\keyword{internal}
