% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode_beta}
\alias{decode_beta}
\title{Beta value decoder}
\usage{
decode_beta(encoded)
}
\arguments{
\item{encoded}{The bit-packed beta and cov Int}
}
\value{
The beta value
}
\description{
Beta value decoder
}
