% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode_cov}
\alias{decode_cov}
\title{Coverage value decoder}
\usage{
decode_cov(encoded)
}
\arguments{
\item{encoded}{The bit-packed beta and cov Int}
}
\value{
The coverage value
}
\description{
Coverage value decoder
}
