% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{.get_threads}
\alias{.get_threads}
\title{Get the number of threads to use in functions from param or option}
\usage{
.get_threads(nthreads)
}
\description{
Get the number of threads to use in functions from param or option
}
\keyword{internal}
