% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_regions.R
\name{fsummarize}
\alias{fsummarize}
\title{Read WGBS BED file and get summary statistics}
\usage{
fsummarize(bedfile, aligner, merged)
}
\arguments{
\item{bedfile}{Path to the BED file}

\item{aligner}{The aligner used to produce the BED files - one of "biscuit",}

\item{merged}{Whether the input strands have been merged/collapsed
"bismark", "bsbolt".}
}
\value{
A data.table
}
\description{
Read WGBS BED file and get summary statistics
}
\keyword{internal}
