% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_beta_sparse}
\alias{get_beta_sparse}
\title{Get beta values from bit-packed sparse matrix}
\usage{
get_beta_sparse(m)
}
\arguments{
\item{m}{The bit-packed dgcMatrix}
}
\value{
The beta value
}
\description{
Get beta values from bit-packed sparse matrix
}
\keyword{internal}
