% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_cov_sparse}
\alias{get_cov_sparse}
\title{Get coverage values from bit-packed sparse matrix}
\usage{
get_cov_sparse(m)
}
\description{
Get coverage values from bit-packed sparse matrix
}
\keyword{internal}
