% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{get_df_input_regions}
\alias{get_df_input_regions}
\title{Return data.frame from string region inputs to write to disk}
\usage{
get_df_input_regions(regions)
}
\description{
Return data.frame from string region inputs to write to disk
}
\keyword{internal}
