% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{get_df_string}
\alias{get_df_string}
\title{DataFrame to region strings}
\usage{
get_df_string(regions_df, feature_col = NULL)
}
\arguments{
\item{regions_df}{A data frame with "chr", "start" and "end" columns}

\item{feature_col}{The data frame column to use as the names of the output string vector}
}
\value{
A character vector
}
\description{
Convert DataFrame to a vector of strings. Set feature names in a "name" column
}
\examples{
(df <- data.frame(chr = c("chr1", "chr2"), start = c(1, 5), end = c(4, 10)))
get_df_string(df)
}
