% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_log_level}
\alias{get_log_level}
\title{Get the current log level}
\usage{
get_log_level()
}
\value{
The current logging level as a string
}
\description{
Can handle all of spdlogs levels, but iscream functions only supports
"info" and "debug"
}
