% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_regions.R
\name{get_named_regions}
\alias{get_named_regions}
\title{Return region strings from GRanges or data.frame region inputs}
\usage{
get_named_regions(regions, feature_col)
}
\description{
Return region strings from GRanges or data.frame region inputs
}
\keyword{internal}
