% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_omp_threads}
\alias{get_omp_threads}
\title{Get the numebr of available threads}
\usage{
get_omp_threads(verbose)
}
\arguments{
\item{verbose}{Whether to be verbose on available omp threads}
}
\description{
Get the numebr of available threads
}
\examples{
get_omp_threads(verbose = T)
}
