% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_regions.R
\name{get_regions_as_df}
\alias{get_regions_as_df}
\title{Return the input regions as a data frame so the output can have positions as columns}
\usage{
get_regions_as_df(regions, regions_str)
}
\description{
Return the input regions as a data frame so the output can have positions as columns
}
\keyword{internal}
