% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{get_string_input_regions}
\alias{get_string_input_regions}
\title{Return region strings from GRanges or data.frame region inputs}
\usage{
get_string_input_regions(regions, feature_col = NULL)
}
\description{
Return region strings from GRanges or data.frame region inputs
}
\keyword{internal}
