% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sample.R
\name{joiner}
\alias{joiner}
\title{Cpg and sample joiner}
\usage{
joiner(cpg_table, sample_dt, sample_name)
}
\arguments{
\item{cpg_table}{The table with CpG loci and any existing samples}

\item{sample_dt}{The sample data table from read_sample}

\item{sample_name}{name of the sample getting joined}
}
\value{
The cpg_table with the sample_table tacked on
}
\description{
Takes a sample data.table and joins it to the CpG index table on the
chromosome and start columns
}
