% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_bsseq}
\alias{make_bsseq}
\title{Make a Bsseq object}
\usage{
make_bsseq(bedfiles, regions, print)
}
\arguments{
\item{bedfiles}{A vector of bedfiles}

\item{regions}{A vector of regions}

\item{print}{Whether to print the Cov and M matrices}
}
\description{
Make a Bsseq object
}
