% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bsseq.R
\name{make_bsseq_from_sparse}
\alias{make_bsseq_from_sparse}
\title{Make a BSSeq object}
\usage{
make_bsseq_from_sparse(sp, sample_list)
}
\arguments{
\item{encoded_matrix}{A sparse matrix of CpG loci and samples with encoded beta values}
}
\value{
A BSSeq object
}
\description{
Create a BSSeq object using a encoded sparse methylation matrix
}
