% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bsseq.R
\name{make_bsseq_lm}
\alias{make_bsseq_lm}
\title{Make a BSSeq object using less memory}
\usage{
make_bsseq_lm(
  sparse_mat,
  sample_list,
  cov_filename = "cov.mtx.gz",
  m_filename = "M.mtx.gz"
)
}
\arguments{
\item{sparse_mat}{A sparse matrix/long table of CpG loci and samples with beta and cov values}

\item{sample_list}{A list of sample names for the bed files}

\item{cov_filename}{Filename to write the coverage matrix market file to}

\item{m_filename}{Filename to write the M matrix market file to}
}
\value{
A BSSeq object
}
\description{
Create a BSSeq object using a encoded sparse methylation matrix. This method
uses less memory than \code{make_bsseq_hm} since it uses sparse coverage and
methylation matrices when creating the BSseq object. It is however slower in
the BSseq object creation.
}
