% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoders.R
\name{make_coverage_mat}
\alias{make_coverage_mat}
\title{Cov matrix}
\usage{
make_coverage_mat(encoded_matrix)
}
\arguments{
\item{encoded_matrix}{A matrix CpG loci and samples with encoded beta values}
}
\value{
A data.table of coverage values for each sample
}
\description{
Make a cov matrix with methylation matrix. Wrapper for \code{decoder}.
}
