% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoders.R
\name{make_m_mat}
\alias{make_m_mat}
\title{M value matrix}
\usage{
make_m_mat(encoded_matrix)
}
\arguments{
\item{encoded_matrix}{A matrix CpG loci and samples with encoded beta values}
}
\value{
A data.table of M values for each sample
}
\description{
Make a beta matrix with methylation matrix. Wrapper for \code{decoder}.
}
\keyword{experimental}
