% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.r
\name{make_sparse_mat}
\alias{make_sparse_mat}
\title{Sparse Methylation matrix}
\usage{
make_sparse_mat(cpg_bed_file, sample_list, sample_path)
}
\arguments{
\item{cpg_bed_file}{The gzipped bed file with CpG loci. This is used to make the matrix index}

\item{sample_list}{A list of sample names for the bed files}

\item{sample_path}{Path to the directory containing sample bed files}
}
\description{
Make a methylation matrix with given scWGBS sample bed files
}
\examples{
sample_list <- c("sample1", "sample2", "sample3")
sample_matrix <- make_meth_mat(sample_list, "./pileup", "./data/cpgs.bed.gz")
}
