% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sparse_mat.R
\name{make_sparse_mat_bs}
\alias{make_sparse_mat_bs}
\title{BSSeq Sparse Methylation matrix}
\usage{
make_sparse_mat_bs(
  cpg_bed_file,
  sample_list,
  sample_path,
  file_ext = ".bed.gz"
)
}
\arguments{
\item{cpg_bed_file}{The gzipped bed file with CpG loci. This is used to make the matrix index}

\item{sample_list}{A list of sample names for the bed files}

\item{sample_path}{Path to the directory containing sample bed files}

\item{file_ext}{The file extension of the sample files}
}
\description{
Make a sparse methylation matrix with given scWGBS sample bed files. The
matrix is sparse in that it stores the cpg locus id, the sample id and the
associated endcoded beta and coverage value. It is not a true sparse matrix
as you would find in the Matrix package and has no functions associated with
it.
}
\keyword{experimental}
