% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mm_writer}
\alias{mm_writer}
\title{Matrix Market writer}
\usage{
mm_writer(DT, row_count, col_count, filename, compress = TRUE)
}
\arguments{
\item{DT}{A data.table with the entries of the matrix to write}

\item{row_count}{The number of rows in the matrix - this is not the same as \code{nrow(DT)}}

\item{col_count}{The number of columns in the matrix - this is not the same as \code{ncol(DT)}}

\item{filename}{The name of the file to write to}

\item{compress}{Whether to write a gzip compressed file or not}
}
\description{
Write a sparse matrix data.table to disk in the Matrix Market format.
}
\details{
Each row of the data.table has to be of the form:
row_num  col_num value
This function requires \code{row_count} and \code{col_count} because the number of
entries can be calculated from the input data.table.
}
\keyword{internal}
