% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{query_chroms}
\alias{query_chroms}
\title{Query the chromosomes or seqnames from a file}
\usage{
query_chroms(fname)
}
\arguments{
\item{fname}{The bedfile name}
}
\value{
A vector of seqnames
}
\description{
Query the chromosomes or seqnames from a file
}
