% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{query_interval}
\alias{query_interval}
\title{Get reads from single genomic regions from multiple tabixed bed file.}
\usage{
query_interval(bedfiles, region)
}
\arguments{
\item{fname}{The name of the bed file - must have a tabix file with the same name and .tbi extension}

\item{regions}{A vector of regions strings of the form "chr:start-end"}
}
\description{
Get reads from single genomic regions from multiple tabixed bed file.
}
