% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{query_regions_from_file}
\alias{query_regions_from_file}
\title{Get list of reads from multiple genomic regions from a tabixed bed file.}
\usage{
query_regions_from_file(fname, regions)
}
\arguments{
\item{fname}{The name of the bed file - must have a tabix file with the same name and .tbi extension}

\item{regions}{A vector of regions strings of the form "chr:start-end"}
}
\description{
Get list of reads from multiple genomic regions from a tabixed bed file.
}
