% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sample.R
\name{read_sample}
\alias{read_sample}
\title{Sample reader}
\usage{
read_sample(sample_name, merged)
}
\arguments{
\item{sample_name}{The name of the sample in the bed file}

\item{merged}{Whether the bed file has CGs merged or unmerged}
}
\value{
The sample as a data.table
}
\description{
Reads scWGBS data from a sample bed file and generates a data.table
}
\examples{
sample1 <- read_sample("sample1", TRUE)
}
