% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sample.R
\name{read_sample_bs}
\alias{read_sample_bs}
\title{BSSeq Sample reader}
\usage{
read_sample_bs(sample_name, merged)
}
\arguments{
\item{sample_name}{The name of the sample in the bed file}

\item{merged}{Whether the bed file has CGs merged or unmerged}
}
\value{
The sample as a data.table
}
\description{
Reads scWGBS data from a sample bed file and generates a data.table
}
\keyword{experimental}
