% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scan_tabix}
\alias{scan_tabix}
\title{Get namde list of reads from a single genomic region from one tabixed bed file}
\usage{
scan_tabix(bedfile, regions)
}
\arguments{
\item{bedfile}{The name of the bed file - must have a corresponding tabix
file with the same name and .tbi extension}

\item{regions}{A vector of region strings in the form "chr:start-end"}
}
\description{
Get namde list of reads from a single genomic region from one tabixed bed file
}
\keyword{internal}
