% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{set_threads}
\alias{set_threads}
\title{Set the number of available threads}
\usage{
set_threads(n_threads)
}
\arguments{
\item{n_threads}{The number of threads to use}
}
\description{
This function sets the 'iscream.thread' option to \code{n_threads}. To
see how many threads are available, run \code{get_omp_threads}.
}
\examples{
set_threads(4)
}
