% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_regions.R
\name{summarize_samples}
\alias{summarize_samples}
\title{Summarize CpGs methylation information over full files}
\usage{
summarize_samples(bedfiles, aligner, merged, nthreads = NULL)
}
\arguments{
\item{bedfiles}{A \code{summarize_regions()} table with chromosomes as regions}

\item{aligner}{The aligner used to produce the BED files - one of "biscuit",
"bismark", "bsbolt".}

\item{merged}{Whether the input strands have been merged/collapsed}

\item{nthreads}{Set number of threads to use overriding the
\code{"iscream.threads"} option. See \code{?set_threads} for more information.}
}
\value{
A data.table
}
\description{
Summarize CpGs methylation information over full files
}
\examples{
bedfiles <- system.file("extdata", package = "iscream") |>
  list.files(pattern = "[a|b|c|d].bed.gz$", full.names = TRUE)

summary <- summarize_samples(bedfiles, merged = FALSE, aligner = "biscuit")
}
