% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_regions.R
\name{summarize_to_sample}
\alias{summarize_to_sample}
\title{Summarize CpGs methylation information over cells}
\usage{
summarize_to_sample(chr_summary)
}
\arguments{
\item{chr_summary}{A \code{summarize_regions()} table with chromosomes as regions}
}
\value{
A data.table
}
\description{
Get average methylation information for the full sample
}
\details{
From https://stats.stackexchange.com/a/389925

\eqn{\mu = \frac{\sum_{n=1}^{n} \mu_{\text{chr}} * n_{\text{chr}}}{\sum_{n=1}^{n} n_{\text{chr}}}}

\eqn{\text{Total CpGs = }N = \sum\limits_{c=1}^{C} n}

\eqn{\text{Sample Mean = }\mu = \frac{\sum\limits_{c=1}^{C} \mu_{\text{c}} n}{N}}

\eqn{\text{Sample Variance} = \sigma^2 = \frac{\sum\limits_{\text{c}=1}^C n(\mu_\text{c}-\mu)^2+\sum\limits_{\text{c}=1}^C n \sigma^2_\text{c}}{N}}

\eqn{\text{Sample stddev} = \sigma = \sqrt{\sigma^2}}
}
\examples{
bedfiles <- system.file("extdata", package = "iscream") |>
  list.files(pattern = "[a|b|c|d].bed.gz$", full.names = TRUE)

chr_summary <- summarize_regions(bedfiles, query_chroms(bedfiles))
summarize_to_sample(chr_summary)
}
