% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{validate_log_level}
\alias{validate_log_level}
\title{Validate provided logging level}
\usage{
validate_log_level(level = get_log_level(), n_threads)
}
\arguments{
\item{level}{The logging level to validate}

\item{n_threads}{The number of threads that the next iscream function call
will use}
}
\value{
None; sets the log level to the provide level
}
\description{
Only "info" and "debug" are currently supported, with "debug" only supported
when using 1 thread
}
\examples{
set_log_level("info")
}
\keyword{internal}
