% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_files.R
\name{validate_region}
\alias{validate_region}
\title{Validate a region string}
\usage{
validate_region(region)
}
\arguments{
\item{region}{A vector of genomic regions}
}
\value{
TRUE if region is valid
}
\description{
CAUTION: very basic validation is done here, checking only that there are
one or three elements and that the last two are integers and the first is
a valid integer or among "X", "Y", "M", optionally prefixed with 'chr'
}
\keyword{internal}
