% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_regions.R
\name{validate_summary_function}
\alias{validate_summary_function}
\title{Return functions to use if the input summarizing functions are valid}
\usage{
validate_summary_function(fun, supported_funcs)
}
\description{
Return functions to use if the input summarizing functions are valid
}
\keyword{internal}
