% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vdecoder}
\alias{vdecoder}
\title{Vector Decoder}
\usage{
vdecoder(encoded, measure)
}
\arguments{
\item{encoded}{An encoded sample column from a data.table methylation matrix of CpG loci by samples}

\item{measure}{The required decoded value: 1 for beta, 2 for coverage and 3 for M}
}
\value{
An IntegerVector of unpacked beta, coverage or M values
}
\description{
This function unpacks the encoded \eqn{\beta} and coverage values from the
encoded column of a methylation matrix.
}
\keyword{experimental}
