% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_aligner_or_stop}
\alias{verify_aligner_or_stop}
\title{Validate provided aligner}
\usage{
verify_aligner_or_stop(aligner)
}
\arguments{
\item{aligner}{The input alinger}
}
\value{
true; quits if the input is not among supported_aligners
}
\description{
Only "biscuit", "bismark", and "bsbolt" are currently supported
}
\keyword{internal}
