% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_files.R
\name{verify_regions_or_stop}
\alias{verify_regions_or_stop}
\title{Verify that regions are valid}
\usage{
verify_regions_or_stop(regions)
}
\arguments{
\item{regions}{A vector of genomic regions}
}
\value{
TRUE if all input regions start with 'chr' FALSE if not
}
\description{
Verify that regions are valid
}
\keyword{internal}
