% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpg_apply}
\alias{cpg_apply}
\title{Apply a function over CpGs within features}
\usage{
cpg_apply(bedfiles, regions, fun, mval, region_rownames = FALSE, nthreads = 1L)
}
\arguments{
\item{bedfiles}{A vector of bedfile paths}

\item{regions}{A vector of genomic regions}

\item{fun}{One of the supported functions to apply over the CpGs in the
regions: \code{"aggregate"}, \code{"average"}.}

\item{mval}{Calculates M values when TRUE, use beta values when FALSE}

\item{region_rownames}{Whether to set rownames to the regions strings}

\item{nthreads}{Number of cores to use. See details.}
}
\description{
Apply a function over CpGs within features
}
\details{
The optimal number of threads depends on the number of bedfiles, but is set
to half the available OpenMP cores. See \code{?get_threads} for more details. It
can be manaully set with \code{set_threads()}.
}
