% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cpp_query_all}
\alias{Cpp_query_all}
\title{Query all CpG info into M and coverage matrices}
\usage{
Cpp_query_all(
  bedfiles,
  regions,
  aligner,
  valInd,
  merged,
  sparse,
  prealloc,
  nthreads
)
}
\arguments{
\item{bedfiles}{A vector of bedfiles}

\item{regions}{A vector of regions}

\item{aligner}{The aligner used to make the WGBS BED files, only for
\code{make_bsseq_mat}}

\item{valInd}{The index of the data column needed for the matrix, for \code{make_mat}}

\item{merged}{Whether the input strands have been merged/collapsed}

\item{prealloc}{The number of rows to initialize the matrices with}

\item{nthreads}{Set number of threads to use overriding the
\code{"iscream.threads"} option. See \code{?set_threads} for more information.}
}
\description{
Query all CpG info into M and coverage matrices
}
\keyword{internal}
