% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoders.R
\name{decoder}
\alias{decoder}
\title{Get decoded values}
\usage{
decoder(encoded_matrix, measure, sparse = TRUE)
}
\arguments{
\item{encoded_matrix}{The matrix with bit-packed beta and cov values}

\item{measure}{The required decoded value: 1 for beta, 2 for coverage and 3 for M}

\item{sparse}{Whether the matrix is a sparse matrix object}
}
\value{
A data.table of the beta, coverage or M value by sample
}
\description{
Get the decoded values from an encoded data.table methylation matrix
}
\keyword{experimental}
