% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_df_string}
\alias{get_df_string}
\title{DataFrame to region strings}
\usage{
get_df_string(regions_df)
}
\arguments{
\item{regions_df}{A data frame with "chr", "start" and "end" columns}
}
\value{
A character vector
}
\description{
Convert DataFrame to a vector of strings. Set feature names in a "name" column
}
