% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_granges_string}
\alias{get_granges_string}
\title{GRanges to region strings}
\usage{
get_granges_string(gr, sep = c(":", "-"))
}
\arguments{
\item{gr}{A GRanges object}

\item{sep}{Vector of separators to use for genomic string. First element is
used to separate chromosome and coordinates, second separator is used to
separate start and end coordinates.}
}
\value{
A character vector
}
\description{
Convert GRanges object to a vector of strings
}
\details{
This function was adapted from
\href{https://github.com/stuart-lab/signac/blob/0b33b1154f9a610897d1efad2c0065081c4e7132/R/utilities.R#L732C1-L756C2}{stuart-lab/signac}.
}
