% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_granges_string}
\alias{get_granges_string}
\title{GRanges to region strings}
\usage{
get_granges_string(gr)
}
\arguments{
\item{gr}{A GRanges object}
}
\value{
A character vector
}
\description{
GRanges to region strings
}
\examples{
if (requireNamespace("GenomicRanges", quietly = TRUE)) {
  get_granges_string(GenomicRanges::GRanges(c("chr1:1-10", "chr2:15-20")))
}
}
