% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_omp_threads}
\alias{get_omp_threads}
\title{Get the number of available threads from OpenMP.}
\usage{
get_omp_threads(verbose)
}
\arguments{
\item{verbose}{Whether to be verbose on available omp threads}
}
\description{
This queries the number of available threads usign OpenMP, but will not
reliably provide an accurate available thread count. To get a more reliable
count that accounts for environment variables and HPC schedulers, use
get_threads()`. This function was pulled from
github.com/rdatatable/data.table
}
\examples{
get_omp_threads(verbose = TRUE)
}
\keyword{internal}
