% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{get_threads}
\alias{get_threads}
\title{Get the number of available threads}
\usage{
get_threads(verbose = FALSE)
}
\arguments{
\item{verbose}{Whether to be verbose on available omp threads}
}
\description{
This function returns the half the number of available OpenMP threads. To
set the number of threads use setDTthreads sets it to half the available
number.
}
\examples{
get_threads(verbose = TRUE)
}
