% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{get_threads}
\alias{get_threads}
\title{Get the number of available threads}
\usage{
get_threads()
}
\value{
A named vector:
\itemize{
\item \code{use_threads} = the number of threads iscream will use
\item \code{opt_set} = whether the option was set by the user
\item \code{avail_threads} = The number of available threads as reported by
\code{parallelly::availableCores}
}
}
\description{
Gets the number of threads iscream is currently set to use, whether the
\code{"iscream.threads"} option is set and how many threads are available for
use. To set the number of threads use \code{set_threads()} or set the
\code{iscream.threads} option in your \verb{~/.Rprofile}. See \code{?set_threads} for more
information.
}
\examples{
get_threads()
}
