% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{htslib_version}
\alias{htslib_version}
\title{Get htslib version and available features}
\usage{
htslib_version()
}
\value{
None
}
\description{
Returns the version of htslib being used by iscream and whether features
such as libdeflate support are available. This information may not always
correspond to the htslib version used during iscream's installation if a
different htslib version is available for linking at runtime.
}
\examples{
htslib_version()

}
