% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.r
\name{make_beta_mat}
\alias{make_beta_mat}
\title{Beta-value matrix}
\usage{
make_beta_mat(encoded_matrix)
}
\arguments{
\item{encoded_matrix}{A matrix CpG loci and samples with encoded beta values}
}
\description{
Make a beta matrix with methylation matrix
}
\examples{
sample_matrix <- make_meth_mat(sample_list, "./pileup", "./data/cpgs.bed.gz")
beta_mat <- make_beta_mat(sample_matrix)
}
