% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bsseq.R
\name{make_bsseq_hm}
\alias{make_bsseq_hm}
\title{Make a BSSeq object}
\usage{
make_bsseq_hm(sp, sample_list)
}
\arguments{
\item{sp}{A sparse matrix/long table of CpG loci and samples with encoded beta values}

\item{sample_list}{A list of sample names for the bed files}
}
\value{
A BSSeq object
}
\description{
Create a BSSeq object using a encoded sparse methylation matrix. This method
uses more memory than \code{make_bsseq_lm} since it uses dense coverage and
methylation matrices when creating the BSSeq object.
}
