% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_mat.R
\name{make_meth_mat}
\alias{make_meth_mat}
\title{Methylation matrix}
\usage{
make_meth_mat(
  sample_list,
  sample_path,
  cpg_bed_file,
  file_ext = ".bed.gz",
  merged = FALSE
)
}
\arguments{
\item{sample_list}{A list of sample names for the bed files}

\item{sample_path}{Path to the directory containing sample bed files}

\item{cpg_bed_file}{The gzipped bed file with CpG loci. This is used to make the matrix index}

\item{file_ext}{The extension of the cpg index and sample files}

\item{merged}{Whether the bed files have CPs merged or unmerged}
}
\description{
Make a methylation matrix with given scWGBS sample bed files
}
\examples{
sample_list <- c("sample1", "sample2", "sample3")
sample_matrix <- make_meth_mat(sample_list, "./pileup", "./data/cpgs.bed.gz")
}
