% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_all.R
\name{query_all}
\alias{query_all}
\title{Make a bsseq object of CpGs from input genomic regions}
\usage{
query_all(bedfiles, regions, nthreads = NULL)
}
\arguments{
\item{bedfiles}{A vector of bedfile paths}

\item{regions}{A vector of genomic regions strings}

\item{nthreads}{Set number of threads to use. Should not be necessary as this
is set by \code{option('iscream.threads')}}
}
\value{
A data.frame
}
\description{
Queries the provided regions and produces M and Coverage matrices that are
wrapped in a BSSeq object
}
\examples{
bedfiles <- system.file("extdata", package = "iscream") |>
  list.files(pattern = "[a|b|c|d].bed.gz$", full.names = TRUE)
# examine the bedfiles
colnames <- c("chr", "start", "end", "beta", "coverage")
lapply(bedfiles, function(i) knitr::kable(data.table::fread(i, col.names = colnames)))

# make a vector of regions
regions <- c("chr1:1-6", "chr1:7-10", "chr1:11-14")
query_all(bedfiles, regions)
}
