% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_all.R
\name{query_all}
\alias{query_all}
\title{Make a bsseq object of CpGs from input genomic regions}
\usage{
query_all(bedfiles, regions, bismark = FALSE, nthreads = NULL)
}
\arguments{
\item{bedfiles}{A vector of bedfile paths}

\item{regions}{A vector of genomic regions strings}

\item{bismark}{Whether the input is a bismark coverage file}

\item{nthreads}{Set number of threads to use overriding the
\code{"iscream.threads"} option. See \code{?set_threads} for more information.}
}
\value{
A data.frame
}
\description{
Queries the provided regions and produces M and Coverage matrices that are
wrapped in a BSSeq object. Parallelized across files using threads from the
\code{"iscream.threads"} option.
}
\examples{
bedfiles <- system.file("extdata", package = "iscream") |>
  list.files(pattern = "[a|b|c|d].bed.gz$", full.names = TRUE)
# examine the bedfiles
colnames <- c("chr", "start", "end", "beta", "coverage")
lapply(bedfiles, function(i) knitr::kable(data.table::fread(i, col.names = colnames)))

# make a vector of regions
regions <- c("chr1:1-6", "chr1:7-10", "chr1:11-14")
query_all(bedfiles, regions)
}
