% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_map.R
\name{region_map}
\alias{region_map}
\title{Run a defined function over genomic regions}
\usage{
region_map(bedfiles, regions, fun = "aggregate", mval = TRUE)
}
\arguments{
\item{regions}{A vector of genomic regions strings}

\item{fun}{Function to apply over the region. See details.}

\item{mval}{Whether to calculate the m value (coverage \eqn{$\times$ beta})
when applying the function.}

\item{befiles}{A vector of bedfile paths}
}
\value{
A data.frame
}
\description{
Run a function on the CpGs in bedfiles across genomic regions. Currently
supported functions are aggregate, average, and ...
}
\details{
Available function:
\itemize{
\item "aggregate"\verb{sums the values in the region with aggregated beta values if}mval =\verb{FALSE and aggregated M values if}mval =` TRUE
\item "average"\verb{averages the values in the region with average beta values if}mval =\verb{FALSE and average M values if}mval =` TRUE
}
}
