% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_meth_regions.R
\name{summarize_meth_regions}
\alias{summarize_meth_regions}
\title{Summarize methylation information over genomic regions}
\usage{
summarize_meth_regions(
  bedfiles,
  regions,
  fun = "all",
  aligner = "biscuit",
  feature_col = NULL,
  mval = TRUE,
  nthreads = NULL
)
}
\arguments{
\item{bedfiles}{A vector of BED file paths}

\item{regions}{A vector, data frame or GenomicRanges of genomic regions. See
details.}

\item{fun}{Function(s) to apply over the region. See details.}

\item{aligner}{The aligner used to produce the BED files - one of "biscuit",
"bismark", "bsbolt".}

\item{feature_col}{Column name of the input \code{regions} data frame containing
a name for each genomic region. Set only if the using a data frame as the
input regions format. See details.}

\item{mval}{Whether to calculate the M value (coverage \eqn{\times \beta})
or use the beta value when applying the function.}

\item{nthreads}{Set number of threads to use overriding the
\code{"iscream.threads"} option. See \code{?set_threads} for more information.}
}
\value{
A data.table
}
\description{
Run summarizing functions on the CpG/CpH loci in BED files across genomic
regions. Parallelized across files using threads from the
\code{"iscream.threads"} option.
}
\section{Supported functions}{
\itemize{
\item Sum: \code{"sum"}
\item Mean: \code{"mean"}
\item Median: \code{"median"}
\item Mode: \code{"mode"}
\item Anti-mode: \code{"antimode"}
\item Standard deviation: \code{"stddev"}
\item Variance: \code{"variance"}
\item Minimum: \code{"min"}
\item Maximum: \code{"max"}
\item Minimum of absolute values: \code{"absmin"}
\item Maximum of absolute values: \code{"absmax"}
\item Range: \code{"range"}
\item First element: \code{"first"}
\item Last element: \code{"last"}
\item No. of records in the region: \code{"count"}
\item No. of records in the region with unique data values: \code{"count_distinct"}
}

Most summarizing computations are backed by the Armadillo library. See
\url{https://arma.sourceforge.net/docs.html#stats_fns} for futher details on the
supported functions
}

\section{Using feature identifiers}{
\code{regions} may be a string vector in the form "chr:start-end", a GRanges
object or a data frame with "chr", "start", and "end" columns. If the input
data.frame or GRanges has a column with feature identifiers, like gene names
for a set of gene regions, pass that column's name to \code{feature_col}. If
\code{regions} is a vector, set its \code{names()} to those identifiers. These will be
used to populate a 'feature' column in the summary. See examples.
}

\examples{
# also see examples from ?summarize_regions

bedfiles <- system.file("extdata", package = "iscream") |>
  list.files(pattern = "[a|b|c|d].bed.gz$", full.names = TRUE)

# make a vector of regions
regions <- c("chr1:1-6", "chr1:7-10", "chr1:11-14")
summarize_meth_regions(bedfiles, regions)

# add names to the regions to populate the 'feature' column
names(regions) <- c("gene1", "gene2", "gene3")
summarize_meth_regions(bedfiles, regions, fun = c("mean", "stddev"), mval = FALSE)
summarize_meth_regions(bedfiles, regions, fun = "sum")
}
