% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_regions.R
\name{summarize_regions}
\alias{summarize_regions}
\title{Summarize information over genomic regions from any BED file}
\usage{
summarize_regions(
  bedfiles,
  regions,
  columns,
  col_names = NULL,
  fun = "all",
  feature_col = NULL,
  nthreads = NULL
)
}
\arguments{
\item{bedfiles}{A vector of BED file paths}

\item{regions}{A vector, data frame or GenomicRanges of genomic regions. See
details.}

\item{columns}{A vector of indices of the numeric columns to be summarized}

\item{col_names}{A vector of names to use for \code{columns} in the output}

\item{fun}{Function(s) to apply over the region. See details.}

\item{feature_col}{Column name of the input \code{regions} data frame or GRanges
\code{mcols} containing a name for each genomic region. Set only if the using a
data frame-like object as the input regions format, not for string vectors.
See details.}

\item{nthreads}{Set number of threads to use overriding the
\code{"iscream.threads"} option. See \code{?set_threads} for more information.}
}
\value{
A data.table
}
\description{
Run summarizing functions on BED file records across genomic regions.
Parallelized across files using threads from the \code{"iscream.threads"} option.
}
\section{Supported functions}{
\itemize{
\item Sum: \code{"sum"}
\item Mean: \code{"mean"}
\item Median: \code{"median"}
\item Mode: \code{"mode"}
\item Anti-mode: \code{"antimode"}
\item Sample standard deviation: \code{"stddev"} (\code{sstdev} in \verb{bedtools map})
\item Population standard deviation: \code{"pstddev"} (\code{stdev} in \verb{bedtools map})
\item Variance: \code{"variance"}
\item Minimum: \code{"min"}
\item Maximum: \code{"max"}
\item Minimum of absolute values: \code{"absmin"}
\item Maximum of absolute values: \code{"absmax"}
\item Range: \code{"range"}
\item First element: \code{"first"}
\item Last element: \code{"last"}
\item No. of records in the region: \code{"count"}
\item No. of records in the region with unique data values: \code{"count_distinct"}
}

Most summarizing computations are backed by the Armadillo library. See
\url{https://arma.sourceforge.net/docs.html#stats_fns} for futher details on the
supported functions
}

\section{Using feature identifiers}{
\code{regions} may be a string vector in the form "chr:start-end", a GRanges
object or a data frame with "chr", "start", and "end" columns. If the input
data.frame or GRanges has a column with feature identifiers, like gene names
for a set of gene regions, pass that column's name to \code{feature_col}. If
\code{regions} is a vector, set its \code{names()} to those identifiers. These will be
used to populate a 'feature' column in the summary. See examples.
}

\examples{
bedfiles <- system.file("extdata", package = "iscream") |>
  list.files(pattern = "[a|b|c|d].bed.gz$", full.names = TRUE)
# examine the bedfiles
colnames <- c("chr", "start", "end", "beta", "coverage")
lapply(bedfiles, function(i) knitr::kable(read.table(i, col.names = colnames)))

# make a vector of regions
regions <- c("chr1:1-6", "chr1:7-10", "chr1:11-14")
summarize_regions(bedfiles, regions, columns = c(4, 5), col_names = c("beta", "cov"))

# select functions
summarize_regions(
  bedfiles,
  regions,
  fun = c("mean", "stddev"),
  columns = c(4, 5),
  col_names = c("beta", "cov")
)

# add names to the regions
names(regions) <- c("gene1", "gene2", "gene3")
summarize_regions(
  bedfiles,
  regions,
  fun = "sum",
  columns = 5,
  col_names = "coverage"
)

# using `feature_col`
library(data.table)

# convert string vector to a data.table
regions_df <- data.table::as.data.table(regions) |>
_[, tstrsplit(regions, ":|-", fixed = FALSE, names = c("chr", "start", "end"))] |>
_[, feature := names(regions)][]
regions_df

summarize_regions(
  bedfiles,
  regions_df,
  fun = "sum",
  columns = 5,
  col_names = "coverage",
  feature_col = "feature"
)
}
