% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.R
\name{tabix}
\alias{tabix}
\title{Query lines from a tabixed bedfile}
\usage{
tabix(
  bedfiles,
  regions,
  aligner = "biscuit",
  col.names = NULL,
  raw = FALSE,
  nthreads = NULL
)
}
\arguments{
\item{bedfiles}{The bedfiles to be queried}

\item{regions}{A vector, data frame or GenomicRanges of genomic regions. See
details.}

\item{aligner}{The aligner used to produce the BED files - one of "biscuit",
"bismark", "bsbolt". Will set the result data.table's column names based on
this argument.}

\item{col.names}{A vector of column names for the result data.table. Set if
your bedfile is not from the supported aligners or is a general bedfile.}

\item{raw}{Set true to give a named list of raw strings from the regions in
the style of \code{Rsamtools::scanTabix} instead of a data.table}

\item{nthreads}{Set number of threads to use overriding the
\code{"iscream.threads"} option. See \code{?set_threads} for more information.}
}
\value{
A data.table
}
\description{
Query lines from a tabixed bedfile
}
\details{
\subsection{Query method}{

'\emph{iscream} has two methods to query records from BED files:
\itemize{
\item the \emph{tabix} shell executable: fast since its output can be redirected to a
file (which \code{data.table::fread()} can then read) instead of having to
allocate memory and store it during the query
\item \emph{iscream's} tabix implementation, based on the \emph{tabix} executable using
\emph{htslib}, but slower on large queries since it stores the records as they
are found instead of writing to a file. However it's able to store each
regions records independently instead of in a single file and is used in
\code{query_all()} and \code{summarize_regions()}.
}

When \emph{iscream} is attached, it checks that the \emph{tabix} executable is
available with \code{Sys.which()} and, if available, sets \code{options("tabix.method" = "shell")}. \code{tabix()} then uses the \emph{tabix} executable to make
queries, except if \code{raw = TRUE}. If \emph{tabix} is not found, \emph{iscream} its
tabix implementation. To use only \emph{iscream's} tabix implementation, set
\code{options("tabix.method" = "htslib")}.
}

\subsection{Input region formats}{

The input regions may be string vector in the form "chr:start-end", a
dataframe with "chr", "start" and "end" columns or a GRanges object. If the
input is a GRanges, the output will also be GRanges with any associated
metadata columns (joined onto the result using
\code{GenomicRanges::findOverlaps()})
}
}
\examples{
bedfiles <- system.file("extdata", package = "iscream") |>
  list.files(pattern = "[a|b|c|d].bed.gz$", full.names = TRUE)
regions <- c("chr1:1-6", "chr1:7-10", "chr1:11-14")
tabix(bedfiles[1], regions, col.names = c("chr", "start", "end", "beta", "coverage"))
}
