% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{tabix}
\alias{tabix}
\title{Query lines from a tabixed bedfile}
\usage{
tabix(
  bedfile,
  regions,
  aligner = "biscuit",
  colnames = NULL,
  raw = FALSE,
  nthreads = NULL
)
}
\arguments{
\item{bedfile}{The bedfile to be queried}

\item{regions}{A vector of genomic region strings}

\item{aligner}{@param aligner The aligner used to produce the BED files -
one of "biscuit", "bismark", "bsbolt". Will set the result data.table's
column names based on this argument.}

\item{colnames}{A vector of column names for the result data.table. Set if
you your bedfile is not from the supported aligners or is a general bedfile.}

\item{raw}{Set true to give a named list of raw strings from the regions in
the style of \code{Rsamtools::scanTabix} instead of a data.table}

\item{nthreads}{Set number of threads to use overriding the
\code{"iscream.threads"} option. See \code{?set_threads} for more information.}
}
\value{
A data.table
}
\description{
Query lines from a tabixed bedfile
}
\examples{
bedfiles <- system.file("extdata", package = "iscream") |>
  list.files(pattern = "[a|b|c|d].bed.gz$", full.names = TRUE)
regions <- c("chr1:1-6", "chr1:7-10", "chr1:11-14")
tabix(bedfiles[1], regions, colnames = c("chr", "start", "end", "beta", "coverage"))
}
