% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vdouble_decoder}
\alias{vdouble_decoder}
\title{Double Vector Decoder
This function unpacks the encoded \eqn{\beta} and coverage values from the
encoded column of a methylation matrix.}
\usage{
vdouble_decoder(encoded, measure)
}
\arguments{
\item{encoded}{An encoded sample column from a data.table methylation matrix of CpG loci by samples}

\item{measure}{The required decoded value: 1 for beta, 2 for coverage and 3 for M}
}
\description{
Double Vector Decoder
This function unpacks the encoded \eqn{\beta} and coverage values from the
encoded column of a methylation matrix.
}
