% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vencoder}
\alias{vencoder}
\title{Vector Encoder}
\usage{
vencoder(beta_col, cov_col)
}
\arguments{
\item{beta_col}{Name of column containing beta values}

\item{cov_col}{Name of column containing coverage values}
}
\description{
Bit-packs a vector of beta and coverage values
}
