% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_files.R
\name{verify_files_or_stop}
\alias{verify_files_or_stop}
\title{Verify that bedfiles are tabixed}
\usage{
verify_files_or_stop(bedfiles, verify_tabix = T)
}
\arguments{
\item{bedfiles}{A vector of bedfile paths}
}
\value{
TRUE if all input bedfiles have an associated tabix index file.
FALSE if not
}
\description{
Verify that bedfiles are tabixed
}
